Remove java cross-comp options, include debuginfo
When compiling Java part with JDK 11 (or later), the current javac
options `-source 1.8` and `-tatget 1.8` are ineffective. Resulting
bytecode format is compatible with JDK 8, but cannot be run on JDk 8
because it is compiled against the bootstrap classpath of JDK 11 and
depends on methods not available in JDK 8.

This change removes these cross-compilation flags. It is assumed, that
official release binaries are always built with JDK 8. For
users/packagers, who is going to build the driver from source, a
CMake configure-time warning about JDK 8 compatibility is added.

Additionally Java debuginfo is added to the JAR (`-g` flag). It does
not affect the performance and the JAR size increase is minor (less
than 10KB).

Fixes: #143