Support typed configuration options
This change adds support for all primitive types (and additionally for
a list of strings) for configuration values that can be passed to
`DriverManager.getConnection(url, conf)`.

To be able to handle config in a separate compilation unit, cached
references and common utilities are moved to separate files. There are
a few more check added to the Java references collection code, but it
still kept largely the same.

Additionally the `file_system` field of the `DBConfig` is filled
manually, it was necessary because this field is used for checks when
incoming file-system related options are set.

Testing: new test added that sets typed config values and checks that
they are set with `current_setting()`.

Fixes: #137