Modify UUID when passing it as a bind variable
When parsing an UUID from a string literal DuckDB modifies it with
flipping the most significant bit in `UUID::FromString` for the
following reason:

"Flip the first bit to make `order by uuid` same as `order by
uuid::varchar`"

When an UUID is read from a result set, JDBC driver flips the MSB again
to get back the original UUID in Java.

This bit-flipping logic was missing when an UUID was passed from Java
as a bind variable. The proposed change adds the same logic when
passing bind variables.

Testing: existing test is improved to cover both string literal and a
bind variable.

Fixes: #147