Do not build jemalloc on Windows
The jemalloc extention was recently added to JDBC builds in #133.
The logic for its selection in `vendor.py` does not look fully correct
to me, but both Linux and Mac are passing on CI, only Windows is
breaking, so this change only tries to unbreak Windows builds, and
the jemalloc selection logic can be improved in subsequent PRs.