Only export JNI and C-API symbols
Currently libduckdb_java.so shared lib exports all symbols. This is not
good for a library that needs to be loaded into JVM process.

Proposed change makes the shared lib to only export the JNI and C-API
symbols. C-API symbols are exported with the intention to be used by
C-API-based extensions.

Testing: no functional changes, no new tests.

Fixes: #102