Result set Timestamp handling improvements
DuckDB supports both `TIMESTAMP WITHOUT TIME ZONE` and
`TIMESTAMP WITH TIME ZONE`. When gettting Java `Timestamp` from a
result set, user can optionally request desired time zone by specifying
a `Calendar` instance.

Thus we are getting the following invariants:

1. `TIMESTAMP WITH TIME ZONE`, no `Calendar`: returns
`Timestamp` in a system-default (JVM) time zone that represents the
same moment in time as the DB's `TIMESTAMP WITH TIME ZONE`. For example
with a system-default time zone being `EET` (`UTC+2`) the timestamp with
`12:00` time and `CET` (`UTC+1`) DB time zone will be returned with
`13:00` time (in `EET`).

2. `TIMESTAMP WITH TIME ZONE`, user-specified Calendar: the same logic
is used as in point `1.`, just the time-zone from the user-specified
Calendar is used instead of the system-default time zone.

3. `TIMESTAMP WITHOUT TIME ZONE`, no `Calendar`: returns the timestamp
that has the same numeric values of all components as the DB's
`TIMESTAMP WITHOUT TIME ZONE`. For example the timestamp with `12:00`
time will be returned with the same `12:00` time in any system-defaut
time zone.

4. `TIMESTAMP WITHOUT TIME ZONE`, user-specified `Calendar`: this case
is tricky because user is requesting to adjust the timestamp for the
specified timezone, but DB's timestamp does not denote any time zone
that can be used as a starting point for adjustment. In this case we
are using the same logic as Postgres' JDBC driver:

https://github.com/pgjdbc/pgjdbc/blob/d86d7ce19030230f14e573b8095f04721d1d8e35/pgjdbc/src/main/java/org/postgresql/jdbc/TimestampUtils.java#L1606

So we assume that the incoming time stamp belongs to the
system-default time zone. Then we calculating the offset between the
system-default and the user-requested time zones that is effective at
the moment of time defined by our timestamp in a system-default time
zone. Note, that such offset can be different depending on the time of
year due to DST. And then shifting the timestamp by this offset - this
can seen as a "reverse" shift comparing to points `1.` and `2.`. For
example, with a system-default time zone being `EET` (`UTC+2`) and
user-requested time zone being `CET` (`UTC+1`), the timestamp with time
`12:00` will be shifted by the difference between `EET` and `CET` and
returned as `13:00` (in `CET`).

Additionally the logic added for getting date-only and time-only values
from the `TIMESTAMP_*` fields is added. It is done by getting the Java
`Timestamp` first (using the invariants described above) and then
getting date-only or time-only part of this timestamp.

Testing: existing tests adjusted to match the new time zone handling
logic; new test is added for getting `Date` and `Time` from `TIMESTAMP`.

Fixes: #146