Close pending statements on connection close
When connection is closed without closing its statements, the JDBC spec
does NOT require to close such statements automatically. But pending
statements are causing problems (see #101) and having them auto-closed
seems to be a reasonable expectation from a client point of view.

This change adds tracking and auto-closing of statements on connection
close. It also adds/improves synchronizaton on closing connections/
statements/result sets. Straighforward sync on current object is used
deliberately instead of other concurrency methods to keep the
synchronization impl as simple as possible.

Testing: new tests are added to check auto-closing of result sets
and statements.

Fixes: #101