Use engine commit hash when vendoring sources
This PR brings duckdb/duckdb-odbc#102 changes to JDBC repo. The
intention is to only run `Vendor.yml` workflow on a dispatch call from
the mainline DuckDB after the CI runs are passed and extensions are
built. On DuckDB engine side the workflow dispatch call is being added
in duckdb/duckdb#17099 .