Synchronize access to Connection, Statement and ResultSet
This change adds synchronization between accessing and deleting
underlying native objects for `Connection`, `Statement` and
`ResultSet`. All synchronization is done in JNI part, Java-level
synchronization is removed from a few places where it was used.
`volatile` fields are added when checking whether the object is closed
or not.

`Connection`'s underlying native object maintains a list of `Statement`s
currently open on this `Connection`. These statements are closed when
the connection is closed. Running queries are cancelled (interrupted)
automatically when the `Connection` is closed.

Note: `Statement.close()` is blocked if a long query, that was started
from this statement, is still running. `Statement.cancel()` must be
called manually before calling `close()` to perform the closing
promptly. This cannot be done automatically, because `cancel()` is
implemented in the engine as a `Connection`-level operation, thus
calling `cancel()` on a `Statement` can interrupt the query running
on another `Statement` on the same `Connection`.

Synchronization for `Appender` is going to be added in a separate PR.

Testing: new tests added for various sequential and concurrent closure
scenarios.

Fixes: #101