Ignore path option passed by Spark
Apache Spark passes the `path` option when a SELECT on a JDBC DataSource
table is performed. It is the internal Spark option and is likely
passed by mistake, so we need to ignore it to allow the connection
to be established.

Testing: new test added that checks that the `path` option is ognored.

Fixes: #55