Make ResultSet getters case insensitive
This change brings in the #43 PR with additional changes to the
`Array` result set requested by the reviewer in #43.

For the `Array` result set the column names `INDEX` and `VALUE` are
taken from the implementation in Postgres' `pgjdbc` driver.

Testing: test from #43 is included, test for `Array` result set is
enhanced with additional checks.

Fixes: #40

Co-authored-by: jghoman <jghoman@gmail.com>