Fix builds by adding missing types.cpp to CMake
In #168 `types.cpp` file was added to `CMakeLists.txt`, but, by mistake
it was not added to `CMakeLists.txt.in` that is used to generate the
`CMakeLists.txt` file on periodic engine sources import.

This change adds `types.cpp` to both `CMakeLists.txt.in` and
`CMakeLists.txt` to fix the build.