Expose query progress on statement
This change exposes the `duckdb_query_progress` native call as
`getQueryProgress()` method on the statement. Similar to the
`Statement#cancel()` call this new method also expected to be called
when `execute()` is running and also works on the connection (not on
the statement itself).

Note, the local option `SET enable_progress_bar = true` needs to be set
to be able to fetch actual progress data.

Testing: new test is added that checks that `getQueryProgress()`
returns valid values.