Support connection properties in URL
This change adds support for specifying connection properties as part
of the connection string itself. It is intended to be used in cases,
when `Properties` map is passed by some tool that does not allow direct
access to it, and some of these `Properties` values need to be
overridden.

Properties in URL are accepted in `key=value` format, `;` is used as a
separator, empty properties are ignored.

URL example: `jdbc:duckdb:;allow_community_extensions=true;;allow_unsigned_extensions = false;`.

Testing: new test added that checks that properties in URL take
preference.