Fix deadlock when fetching next chunk (1.3)
This is a backport of the PR #242 to `v1.3-ossivalis` stable branch.

When connection is being closed, it closes all statements and all result
sets on these statements. All this is done while holding connection
lock.

When fetching next chunk, result set needs to take both result set and
connection locks. Because of the wrong order of taking these locks,
concurrent `close()` call on result set, initiated from
`Connection#close()` was causing a deadlock.

This change reorders locks in `ResultSet#fetchChunk()` to fix this.

Testing: new test added that reproduces the deadlock reliably (may need
to raise the number of iterations).

Fixes: #241