Implement getPropertyInfo method for config (1.3)
This is a backport of the PR #244 to `v1.3-ossivalis` stable branch.

`Driver#getPropertyInfo()` method allows external tools to list
configuration options suppored by the DB. This method is pretty obscure
and unlikely to be useful with many client tools in practice. But a
full list of supported options is required anyway to implement a fix
to issue #232, and it makes sense to use `getPropertyInfo` for that as
a single method for this functionality.

Testing: new test added.