Do not allow to cancel other running statements
`Statement#cancel()` call works on connection level. So currently it is
well possible to run query in one statement, and then interrupt this
query calling cancel on another statement.

This change adds a check that cancellation can only be performed if
the query on current statement is still running. Otherwise
`stmt.cancel()` call is a no-op.

Testing: new test added that checks that other statement cannot be
cancelled.