Add support for query timeouts (1.3)
This is a backport of the PR #247 to `v1.3-ossivalis` stable branch.

This change implements `Statement#setQueryTimeout()` method. It is
implemented by scheduling a background task and calling
`Statement#cancel()` when timeout expires.

Timeouted statement has the same behaviour as it would be if cancelled
manually - `SQLException` is thrown and the statement is closed.

Timeout is applied for all `execute*` calls. For `executeBatch()` it is
applied separately for every single query in a batch.

Testing: new test added.

Fixes: #212