Allow to ignore unsupported connection opts (1.3)
This is a backport of the PR #250 to `v1.3-ossivalis` stable branch.

Currently driver performs strict checks on the user-specified connection
options. If stray options are passed - exception is thrown on connect.

This is considered to be a good feature, to prevent accidental typos in
option names. But in some cases user does not have full control over the
options passed to driver by other tools.

This change adds new boolean connection property
`jdbc_ignore_unsupported_options`, when it is enabled - all unsupported
connection options are ignored.

When used with external tools, it is intended to be set in connection
string like this:

```
jdbc:duckdb:;jdbc_ignore_unsupported_options=true;
```

Testing: new test added.

Fixes: #232