Fix executeBatch in auto-commit mode (1.3)
This is a backport of the PR #261 to `v1.3-ossivalis` stable branch.

This change fixes the problem when connection is left in inconsistent
state when `executeBatch` throws constraint (or some other) error with
auto-commit enabled.

In #213 `executeBatch` was changed to effectively disable auto-commit
for the whole batch. In this case the rollback on error needs to be
handled inside `executeBatch` too.

Testing: new tests added that cover constrain violation with
`executeBatch`.

Fixes: #259