Fix vendoring PRs for stable branches (1.3)
This is a backport of the PR #262 to `v1.3-ossivalis` stable branch.

Nightly import of DuckDB engine sources ("vendoring") is performed by
creating PRs on `vendoring-<branch>` base branches. After the tests are
passed the PR is merged and `vendoring-<branch>` is re-created to be
used in subsequent imports.

This change fixes the vendoring branch re-creation to place it not
always on the `main` branch, but on the correct branch (either `main` or
stable one line `v1.3-ossivalis`).