DuckLake options override fix
With the changes introduced in #276 the `jdbc:duckdb:ducklake:...`
URLs have the special handling: `jdbc_pin_db` and `jdbc_stream_results`
options are applied to them automatically. This behaviour was supposed
to be overridable with URL or `Properties` but by mistake incorrect
method was used for override check. And as DuckLake is still not
available in the `main` branch, there is no test coverage for this and
the problem got into `1.3.1.0` update.

Fixes: #283