Backport #18374 engine fix to v1.3.2 branch
This PR contains a cherry-pick of the duckdb/duckdb#18752 engine PR
applied on top of the `v1.3.2.0` release. `v1.3.2` branch is created for
it so the new release built from this branch will be compatible with
existing `v1.3.2` extensions.

Testing: test case included with the original PR is run manually on
patched JDBC.