Return precision values for non-dec columns (1.4)
This is a backport of the PR #367 to `v1.4-andium` stable branch.

This PR makes `ResultSetMetaData#getPrecision()` to return non-zero
values for integral numbers, floating point, date/time and string
columns.

Unlike the `DECIMAL` type, where actual column precision is returned,
for other types this call is expected to give a hint to a client
software how much space the textual representation of the field is
going to take. Previously returned 0 value appeared to cause problems
in some of such clients.

Values returned are consistent with ones returned by Postgres.

Testing: new test added for all types.

Fixes: #340