Appender: support composite/nested lists and maps
This PR adds support for appending to `MAP` and composite `LIST`
columns, for example `LIST` of `STRUCT`s, `LIST` of `ARRAY`s or
nested `LIST`s.

It should cover virtually any table schema with composite fields, with
the notable exception of `ARRAY`s with more than 2 dimensions (like
`INT[2][3][4]`) - only 2D `ARRAY`s are supported. Higher dimensions
could be possible, but the interface for them appeared to be tough to
generalize. `LIST`s of `ARRAY`s can be used instead, like `INT[2][3][]`.

`LIST`s can be specified as a `Collection` or `Iterator` (with `count`)
of objects.

`MAP`s are specified as `java.util.Map` instances.

`STRUCT` inside a `LIST` can be specified either as
`java.utils.LinkedHashMap` (keys are ignored) or as a `Collection` of
values for `STRUCT` fields.

`UNION` inside a `LIST` can be specified as an instance of
`java.util.AbstractMap.SimpleEntry`, where a `key` is used as a tag to
choose which `UNION` field to append to.

If some required input cases are missed - please raise an issue.

Testing: new tests added to cover various combinations of nesting.