Appender: fix buffer size for nested STRUCTs (1.4)
This is a backport of the PR #439 to `v1.4-andium` stable branch.

This change fixes the buffer size calculation for `STRUCT`s nested
inside `LIST`s and for `MAP`s. It also fixes the validity mask size
calculation for these cases.

Testing: additional tests added with longer `MAP`s (to check nested
`STRUCT`s) and with `NULL` values (to check validity masks).

Fix: #437