Add JPMS automatic module name (1.4)
This is a backport of the PR #454 to `v1.4-andium` stable branch.

This PR adds the name `duckdb.jdbc` as an [automatic JPMS module](https://openjdk.org/projects/jigsaw/spec/sotms/#automatic-modules)
name to the driver JAR. This makes this name, that was previously
derived from the JAR file name, to be stable when the driver is used
from modular applications.

Testing: CI test added that checks the module name.

Fixes: #431