Disable Spatial tests
This PR temporary disables the Spatial extension tests to allow the full
test suite to run on the latest `main` where Spatial is not currently
available.

Additionally it adds a test that uses Excel extension. This test is also
used for `duckdb_jdbc_nolib.jar` checks instead of Spatial tests.