Add jemalloc placeholder on Windows
Due to the way how `generated_extension_loader_package_build.cpp` is
generated in the latest `main` versions of the engine, it expects to
have the same set of linked extensions on all platforms. `jemalloc`
extension is not available on Windows, so the JDBC build fails on
Windows.

This PR adds an emtpy placeholder for `JemallocExtension` class that is
expected by `generated_extension_loader_package_build.cpp` to fix the
compilation on Windows.